result 对象修改 1.3.8

+++ src/main/java/com/fxipp/kratos/result/ErrorCode.java
@@ -4,7 +4,7 @@ package com.fxipp.kratos.result;
  * 封装API的错误码
  */
 public interface ErrorCode {
-    Long getCode();
+    Integer getCode();
 
     String getMessage();
 }
